Match Menu component naming on the upstream lib
