+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Changes
 
+## 0.0.5
+
+* Add MenuItem, MenuHeader, MenuMenu and Menu components
+
 ## 0.0.4
 
 * Use react.common
-* Add MenuItem, MenuHeader, MenuMenu and Menu components
 
 ## 0.0.3
