+++ facade/src/test/scala/react/semanticui/collections/menu/MenuHeaderTests.scala
@@ -6,14 +6,14 @@ import japgolly.scalajs.react.test._
 object MenuHeaderTests extends TestSuite {
   val tests = Tests {
     'render - {
-      val menuItem = MenuHeader(MenuHeader.props())
+      val menuItem = Menu.Header(Menu.Header.props())
       ReactTestUtils.withNewBodyElement { mountNode =>
         menuItem.renderIntoDOM(mountNode)
         assert(mountNode.outerHTML == """<div><div class="header"></div></div>""")
       }
     }
     'renderAs - {
-      val menuItem = MenuHeader(MenuHeader.props(as = "a"))
+      val menuItem = Menu.Header(Menu.Header.props(as = "a"))
       ReactTestUtils.withNewBodyElement { mountNode =>
         menuItem.renderIntoDOM(mountNode)
         assert(mountNode.outerHTML == """<div><a class="header"></a></div>""")
