+++ facade/src/test/scala/react/semanticui/collections/menu/MenuMenuTests.scala
@@ -6,14 +6,14 @@ import japgolly.scalajs.react.test._
 object MenuMenuTests extends TestSuite {
   val tests = Tests {
     'render - {
-      val menuItem = MenuMenu(MenuMenu.props())
+      val menuItem = Menu.Menu(Menu.Menu.props())
       ReactTestUtils.withNewBodyElement { mountNode =>
         menuItem.renderIntoDOM(mountNode)
         assert(mountNode.outerHTML == """<div><div class="menu"></div></div>""")
       }
     }
     'renderAs - {
-      val menuItem = MenuMenu(MenuMenu.props(as = "a"))
+      val menuItem = Menu.Menu(Menu.Menu.props(as = "a"))
       ReactTestUtils.withNewBodyElement { mountNode =>
         menuItem.renderIntoDOM(mountNode)
         assert(mountNode.outerHTML == """<div><a class="menu"></a></div>""")
